changes to paths... runs locally, not container or tests
