--- express-api/constants/index.ts
@@ -1,8 +1,0 @@
-import networking from './networking';
-import switches from './switches';
-
-const constants = {
-  ...networking,
-  ...switches,
-};
-export default constants;
