+++ express-api/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.ts",
   "scripts": {
-    "dev": "nodemon --exec ts-node -r dotenv/config server.ts dotenv_config_path=../.env",
+    "dev": "nodemon --exec ts-node -r tsconfig-paths/register -r dotenv/config ./src/server.ts dotenv_config_path=../.env",
     "build": "tsc",
     "lint": "eslint './' --ext .ts,.js",
     "lint:fix": "npm run lint -- --fix",
@@ -14,7 +14,7 @@
     "coverage": "jest --coverage",
     "test:unit": "jest --testPathPattern=/tests/unit",
     "test:integration": "jest --testPathPattern=/tests/integration",
-    "swagger": "node ./swagger/swagger.mjs"
+    "swagger": "node ./src/swagger/swagger.mjs"
   },
   "author": "",
   "license": "ISC",
@@ -57,6 +57,7 @@
     "swagger-autogen": "2.23.7",
     "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
+    "tsconfig-paths": "4.2.0",
     "typescript": "5.2.2"
   }
 }
