--- express-api/routes/index.ts
@@ -1,9 +1,0 @@
-import ltsaRouter from './ltsaRouter';
-import healthRouter from './healthRouter';
-
-const router = {
-  healthRouter,
-  ltsaRouter,
-};
-
-export default router;
