+++ express-api/src/constants/index.ts
@@ -1,0 +1,8 @@
+import networking from '@constants/networking';
+import switches from '@constants/switches';
+
+const constants = {
+  ...networking,
+  ...switches,
+};
+export default constants;
