+++ express-api/src/routes/healthRouter.ts
@@ -1,4 +1,4 @@
-import controllers from '../controllers';
+import controllers from '@controllers';
 import express from 'express';
 
 const router = express.Router();
