+++ express-api/src/routes/index.ts
@@ -1,0 +1,9 @@
+import ltsaRouter from '@routes/ltsaRouter';
+import healthRouter from '@routes/healthRouter';
+
+const router = {
+  healthRouter,
+  ltsaRouter,
+};
+
+export default router;
