+++ express-api/src/routes/ltsaRouter.ts
@@ -1,4 +1,4 @@
-import controllers from '../controllers';
+import controllers from '@controllers';
 import express from 'express';
 
 const router = express.Router();
