+++ express-api/tests/integration/middleware/headerHandler.test.ts
@@ -1,4 +1,4 @@
-import app from '../../../express';
+import app from '@express';
 import supertest from 'supertest';
 
 const request = supertest(app);
