+++ express-api/tests/unit/controllers/healthController.test.ts
@@ -1,4 +1,4 @@
-import controllers from '@controllers/index';
+import controllers from '@controllers';
 import { Request, Response } from 'express';
 
 describe('UNIT - Testing controller for /health routes', () => {
